Split utils.rs into fetchable.rs and constants.rs. Fixed tests + apply review
