+++ chain-registry/src/lib.rs
@@ -1,8 +1,9 @@
 pub mod asset_list;
 pub mod chain;
+pub mod constants;
 pub mod error;
+pub mod fetchable;
 pub mod formatter;
 pub mod paths;
 pub mod querier;
 pub mod relayer_config;
-pub mod utils;
