+++ chain-registry/src/paths.rs
@@ -1,5 +1,5 @@
 /// Models for serializing and deserializing IBC path JSON data found in the `_IBC/` directory of the registry repository
-use crate::utils::Fetchable;
+use crate::fetchable::Fetchable;
 use ibc::core::ics24_host::identifier::{ChannelId, ClientId, ConnectionId, PortId};
 use serde::Deserialize;
 use serde::Serialize;
@@ -97,7 +97,7 @@ mod tests {
     use super::*;
     use crate::error::RegistryError;
 
-    use crate::utils::ALL_PATHS;
+    use crate::constants::ALL_PATHS;
 
     #[tokio::test]
     async fn fetch_paths() -> Result<(), RegistryError> {
